services/api-gateway/requirements.txt
fastapi
uvicorn
requests
pyjwt
pytest
pytest-cov
httpx